added title to gif frames

+++ Segmentation/utils/evaluation_utils.py
@@ -204,17 +204,7 @@ def plot_and_eval_3D(model,
     #  break
 
     if not gif_dir == '':
-        print("\n\n\n\n=================")
-        print("checking for ffmpeg...")
-        if not os.path.isfile('./../../../opt/conda/bin/ffmpeg'):
-            print("please 'pip install ffmpeg' to create gif")
-            print("gif not created")
-            
-        else:
-            print("ffmpeg found")
-            pred_evolution_gif(fig, images_gif, save_dir=gif_dir, save=True)
-
-        print("=================\n\n\n\n")
+        pred_evolution_gif(fig, images_gif, save_dir=gif_dir, save=True)
 
 def pred_evolution_gif(fig,
                        frames_list,
@@ -223,30 +213,39 @@ def pred_evolution_gif(fig,
                        save=True,
                        show=False):
 
-    print("creating the gif ...\n")
-    gif = ArtistAnimation(fig, frames_list, interval, repeat=True) # create gif
+    print("\n\n\n\n=================")
+    print("checking for ffmpeg...")
+    if not os.path.isfile('./../../../opt/conda/bin/ffmpeg'):
+        print("please 'pip install ffmpeg' to create gif")
+        print("gif not created")
+        
+    else:
+        print("ffmpeg found")
+        print("creating the gif ...\n")
+        gif = ArtistAnimation(fig, frames_list, interval, repeat=True) # create gif
 
-    if save:
-        if save_dir == '':
-            time = datetime.now().strftime("%Y%m%d-%H%M%S")
-            save_dir = 'results/gif'+ time + '.gif'
+        if save:
+            if save_dir == '':
+                time = datetime.now().strftime("%Y%m%d-%H%M%S")
+                save_dir = 'results/gif'+ time + '.gif'
 
-        plt.rcParams['animation.ffmpeg_path'] = r'//opt//conda//bin//ffmpeg'  # set directory of ffmpeg binary file
-        Writer = animation.writers['ffmpeg']
-        ffmwriter = Writer(fps=1000//interval, metadata=dict(artist='Me'), bitrate=1800) #set the save writer
-        gif.save('results/temp_video.mp4', writer=ffmwriter)
+            plt.rcParams['animation.ffmpeg_path'] = r'//opt//conda//bin//ffmpeg'  # set directory of ffmpeg binary file
+            Writer = animation.writers['ffmpeg']
+            ffmwriter = Writer(fps=1000//interval, metadata=dict(artist='Me'), bitrate=1800) #set the save writer
+            gif.save('results/temp_video.mp4', writer=ffmwriter)
 
-        codeBASH = f"ffmpeg -i 'results/temp_video.mp4' -loop 0 {save_dir}" #convert mp4 to gif
-        os.system(codeBASH)
-        os.remove("results/temp_video.mp4")
+            codeBASH = f"ffmpeg -i 'results/temp_video.mp4' -loop 0 {save_dir}" #convert mp4 to gif
+            os.system(codeBASH)
+            os.remove("results/temp_video.mp4")
 
-        plt.close('all')
+            plt.close('all')
 
-    if show:
-        plt.show()
-        plt.close('all')
-    
-    print('\ndone')
+        if show:
+            plt.show()
+            plt.close('all')
+        
+        print("\n\n=================")
+        print('done\n\n')
 
 def confusion_matrix(trained_model,
                      weights_dir,
